some refactor
