+++ api/grpc/GrpcClient.java
@@ -1,10 +1,9 @@
+import context.Context;
 import io.grpc.Channel;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
 import io.grpc.StatusRuntimeException;
 import keploy.Keploy;
-import keploy.KeployInstance;
-import keploy.context.Context;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import stubs.RegressionServiceGrpc;
@@ -28,13 +27,20 @@ public class GrpcClient {
     private static final Logger logger = LogManager.getLogger("GrpcClient");
     private RegressionServiceGrpc.RegressionServiceBlockingStub blockingStub;
 
+    private ManagedChannel channel;
+
     public GrpcClient() {
-
-    }
-
-    public GrpcClient(Channel channel) {
+        this.channel = ManagedChannelBuilder.forTarget("localhost:8081")
+                // Channels are secure by default (via SSL/TLS). For the example we disable TLS to avoid
+                // needing certificates.
+                .usePlaintext()
+                .build();
         this.blockingStub = RegressionServiceGrpc.newBlockingStub(channel);
     }
+
+//    public GrpcClient(Channel channel) {
+//        this.blockingStub = RegressionServiceGrpc.newBlockingStub(channel);
+//    }
 
 
     public List<Service.getTCSResponse> fetch() {
@@ -270,30 +276,27 @@ public class GrpcClient {
 
     }
 
-
-    public static void main(String[] args) {
-        String target = "localhost:8081";
-        ManagedChannel channel = ManagedChannelBuilder.forTarget(target)
-                // Channels are secure by default (via SSL/TLS). For the example we disable TLS to avoid
-                // needing certificates.
-                .usePlaintext()
-                .build();
-
-        RegressionServiceGrpc.RegressionServiceBlockingStub stub = RegressionServiceGrpc.newBlockingStub(channel);
-
-        Service.startResponse startResponse = stub.start(Service.startRequest.newBuilder().setApp("test").setTotal("1").build());
-        System.out.println(startResponse + "This is the response I am getting !! ");
-        Service.endResponse endResponse = stub.end(Service.endRequest.newBuilder().setStatus("OK").setId("123").build());
-        System.out.println(endResponse + "This is the response I am getting !! ");
-//        Service.TestCase testCase = stub.getTC(Service.getTCRequest.newBuilder().setApp("sample-url-shortener").setId("09a39684-c551-42a9-9b64-3f61815a8662").build());
-//        System.out.println(testCase);
-//       Service.getTCSResponse testcases =  stub.getTCS(Service.getTCSRequest.newBuilder().setApp("sample-url-shortener").setOffset("0").setLimit("25").build());
-//        System.out.println(testcases);
-        GrpcClient k = new GrpcClient(channel);
-        List<Service.getTCSResponse> testCases = k.fetch();
-        System.out.println(testCases);
-        System.out.println(testCases.size());
-    }
-
-
+//    public static void main(String[] args) {
+////        String target = "localhost:8081";
+//        ManagedChannel channel = ManagedChannelBuilder.forTarget("localhost:8081")
+//                // Channels are secure by default (via SSL/TLS). For the example we disable TLS to avoid
+//                // needing certificates.
+//                .usePlaintext()
+//                .build();
+//
+//        RegressionServiceGrpc.RegressionServiceBlockingStub stub = RegressionServiceGrpc.newBlockingStub(channel);
+//
+//        Service.startResponse startResponse = stub.start(Service.startRequest.newBuilder().setApp("test").setTotal("1").build());
+//        System.out.println(startResponse + "This is the response I am getting !! ");
+//        Service.endResponse endResponse = stub.end(Service.endRequest.newBuilder().setStatus("OK").setId("123").build());
+//        System.out.println(endResponse + "This is the response I am getting !! ");
+////        Service.TestCase testCase = stub.getTC(Service.getTCRequest.newBuilder().setApp("sample-url-shortener").setId("09a39684-c551-42a9-9b64-3f61815a8662").build());
+////        System.out.println(testCase);
+////       Service.getTCSResponse testcases =  stub.getTCS(Service.getTCSRequest.newBuilder().setApp("sample-url-shortener").setOffset("0").setLimit("25").build());
+////        System.out.println(testcases);
+//        GrpcClient k = new GrpcClient(channel);
+//        List<Service.getTCSResponse> testCases = k.fetch();
+//        System.out.println(testCases);
+//        System.out.println(testCases.size());
+//    }
 }
